+++ www/components/issues/controllers/issues.controller.js
@@ -1,7 +1,5 @@
 module.exports = function ($scope, IssuesFactory) {
 
-  $scope.issues = IssuesFactory.getIssues();
-  
   IssuesFactory.getIssues().then(function(issues) {
         $scope.issues = issues;
     });
