+++ www/components/map/map.module.js
@@ -3,7 +3,7 @@ var mapModule = angular.module('zonnebloem.map', []);
 mapModule.controller("IssueMapController", require("./controllers/issue.map.controller.js"));
 mapModule.controller('MapController', require('./controllers/map.controller.js'));
 
-mapModule.factory("IssueMapFactory", require("./factories/issue.map.factory.js"));
+mapModule.factory("MapFactory", require("./factories/map.factory.js"));
 mapModule.factory("RoutesWalkedFactory", require("./factories/routesWalked.factory.js"));
 
 mapModule.config(require("./map.routes.js"));
